% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.prepivot.ks.permtest.R
\name{summary.prepivot.ks.permtest}
\alias{summary.prepivot.ks.permtest}
\title{Summarizing Two-sample Goodness-of-fit Permutation Test under Covariate-adaptive Randomization}
\usage{
\method{summary}{prepivot.ks.permtest}(object, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{an object of class \code{"prepivot.ks.permtest"}, the result of calling \code{\link{prepivot.ks.permtest}}}

\item{...}{unused}

\item{digits}{number of digits to display}
}
\value{
\code{summary.prepivot.ks.permtest} returns an object of \link{class} "\code{summary.prepivot.ks.permtest}" which has the following components
 \item{results}{Matrix with the Testing Problem, Sample Sizes, Number of Permutations, Number of Bootstrap samples, Observed test Statistic, Critical value and P-value.}
}
\description{
\code{summary} method for class \code{"prepivot.ks.permtest"}
}
\author{
Maurcio Olivares
}
