% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PT.Khmaladze.fit.R
\name{summary.PT.Khmaladze.fit}
\alias{summary.PT.Khmaladze.fit}
\title{Summarizing Permutation Test for Heterogeneous Treatment Effects with Estimated Nuisance Parameter}
\usage{
\method{summary}{PT.Khmaladze.fit}(object, ..., digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{object}{an object of class \code{"PT.Khmaladze.fit"}, usually a result of a call to \code{\link{PT.Khmaladze.fit}}}

\item{...}{unused}

\item{digits}{number of digits to display}
}
\value{
\code{summary.PT.Khmaladze.fit} returns an object of \link{class} "\code{summary.PT.Khmaladze.fit}" which has the following components
 \item{results}{Matrix with the Testing Problem, Sample Sizes, Number of Permutations, ATE, Test Statistic, Critical value and P-value.}
}
\description{
\code{summary} method for class \code{"PT.Khmaladze.fit"}
}
\author{
Maurcio Olivares Gonzalez

Ignacio Sarmiento Barbieri
}
