% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PT.Khmaladze.MultTest.R
\name{summary.PT.Khmaladze.MultTest}
\alias{summary.PT.Khmaladze.MultTest}
\title{Summarizing Permutation Test for Within-grpup Treatment Effect Heterogeneity in the presence of an Estimated Nuisance Parameter}
\usage{
\method{summary}{PT.Khmaladze.MultTest}(object, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{an object of class \code{"PT.Khmaladze.MultTest"}, usually a result of a call to \code{\link{PT.Khmaladze.MultTest}}}

\item{...}{unused}

\item{digits}{number of digits to display}
}
\value{
\code{summary.PT.Khmaladze.MultTest} returns an object of \link{class} "\code{summary.PT.Khmaladze.MultTest}" which has the following components
 \item{results}{Matrix with the Testing Problem, Number of Permutations for the test and the multiple testing procedure, number of subgroups, (raw) p-values, adjusted p-values, Test Statistic.}
}
\description{
\code{summary} method for class \code{"PT.Khmaladze.fit"}
}
\author{
Maurcio Olivares
}
