% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PTQTE.Khmaladze.fit.R
\name{summary.PTQTE.Khmaladze.fit}
\alias{summary.PTQTE.Khmaladze.fit}
\title{Summarizing Quantile-Based Permutation Test with an Estimated Nuisance Parameter}
\usage{
\method{summary}{PTQTE.Khmaladze.fit}(object, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{an object of class \code{"PTQTE.Khmaladze.fit"}, the result of calling \code{\link{PTQTE.Khmaladze.fit}}}

\item{...}{unused}

\item{digits}{number of digits to display}
}
\value{
\code{PTQTE.Khmaladze.fit} returns an object of \link{class} "\code{PTQTE.Khmaladze.fit}" which has the following components
 \item{results}{Matrix with the Testing Problem, Sample Sizes, Number of Permutations, Observed test Statistic, Binary Rule and Significance Level.}
}
\description{
\code{summary} method for class \code{"PTQTE.Khmaladze.fit"}
}
\author{
Maurcio Olivares
}
