% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RDperm.R
\name{summary.RDperm}
\alias{summary.RDperm}
\title{Summarizing Regression Discontinuity Design Permutation Test}
\usage{
\method{summary}{RDperm}(object, digits = max(3, getOption("digits") -
  3), ...)
}
\arguments{
\item{object}{an object of class \code{"RDperm"}, usually a result of a call to \code{\link{RDperm}}}

\item{digits}{number of digits to display}

\item{...}{unused}
}
\value{
\code{summary.RDperm} returns an object of \link{class} "\code{summary.RDperm}" which has the following components
 \item{results}{Matrix with the Test Statistic, P-values and Q used}
}
\description{
\code{summary} method for class \code{"RDPerm"}
}
\author{
Maurcio Olivares Gonzalez

Ignacio Sarmiento Barbieri
}
