% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RPT.R
\name{summary.RPT}
\alias{summary.RPT}
\title{Summarizing Robust Permutation Tests}
\usage{
\method{summary}{RPT}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"RPT"}, usually a result of a call to \code{\link{RPT}}}

\item{digits}{number of digits to display}

\item{...}{unused}
}
\value{
\code{summary.RPT} returns an object of \link{class} "\code{summary.RPT}" which has the following components:
 \item{T.obs}{Observed test statistic.}
 \item{pvalue}{P-value.}
}
\description{
\code{summary} method for class \code{"RPT"}
}
\author{
Maurcio Olivares Gonzalez

Ignacio Sarmiento Barbieri
}
